+++ module/documents/GSItem.mjs
@@ -9,6 +9,7 @@ export class GSItem extends Item {
     preparedData(){
 
     }
+
     getRollData(){
 
     }